Fixed all the bugs for sampling training. Test passed on all datasets. Tuned hyperpameters for good accuracy.

+++ include/FedTree/common.h
@@ -40,7 +40,7 @@ std::string string_format(const std::string &format, Args ... args) {
 }
 
 //data types
-typedef long double float_type;
+typedef double float_type;
 
 #define EPSILON 1e-9
